Update site layout, content, and configuration
This commit includes several changes to improve the site's usability and content. It modifies some key configuration files, and adds new content and layout files. The main changes are the removal of the English menu, the addition of price-related content and templates, and modifications to improve the website's UI. Logo images and their dimensions have also been updated, as well as enabling and disabling certain site features.